backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
gunicorn
pydantic
